+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Return the tx-id instead of the tx-hash for Bitcoin because it is the canonical way of identifying a transaction.
 
+### Added
+- An order negotiation protocol over HTTP.
+
 ## [0.5.6] - 2019-11-05
 
 ### Fixed